feat: better implementation of custom steamos package

+++ system/modules/gaming.nix
@@ -1,38 +1,68 @@
 {pkgs, ...}:
 
+let
+  steamForGamescope = pkgs.steam.override {
+    extraEnv = {
+      STEAM_EXTRA_COMPAT_TOOLS_PATHS = "${pkgs.proton-ge-bin.steamcompattool}";
+      STEAM_GAMESCOPE_VRR_SUPPORTED = "1";
+    };
+  };
+in
+
 {
   programs.steam = {
     enable = true;
-    gamescopeSession.enable = true;
     extraCompatPackages = with pkgs; [proton-ge-bin];
     package = pkgs.steam.override {
       extraEnv = {
         MANGOHUD = "1";
-        TZ="Europe/Rome";
       };
     };
   };
 
   # SteamDeck UI in Gamescope
   environment.systemPackages = [
-    (pkgs.writeShellScriptBin "steamdeck" ''
+    (pkgs.writeShellScriptBin "steamos" ''
       #!/usr/bin/env bash
-      pkill -9 -x steam
-      pkill -9 -x gamescope-wl
-      sleep 2
 
-      exec env -u MANGOHUD STEAM_GAMESCOPE_VRR_SUPPORTED=1 \
-        ${pkgs.gamescope}/bin/gamescope \
-          --mangoapp \
+      if pgrep -x steam > /dev/null; then
+        ${pkgs.steam}/bin/steam -shutdown
+        
+        for i in {1..10}; do
+          if ! pgrep -x steam > /dev/null; then
+            break
+          fi
+          sleep 1
+        done
+
+        if pgrep -x steam > /dev/null; then
+          pkill -9 -x steam
+          pkill -9 -x gamescope-wl
+          sleep 1
+        fi
+      fi
+
+      exec ${pkgs.gamescope}/bin/gamescope \
           -W 2560 -H 1440 \
           -w 2560 -h 1440 \
           -r 165 \
           -O DP-1 \
+          -f \
           -e \
+          --rt \
+          --mangoapp \
+          --adaptive-sync \
+          --force-grab-cursor \
           --xwayland-count 2 \
-          -- ${pkgs.steam.run}/bin/steam-run ${pkgs.steam}/bin/steam -steamdeck -steamos3
+          -- ${steamForGamescope}/bin/steam -steamdeck -steamos3
     '')
   ];
+
+  # Gamescope
+  programs.gamescope = {
+    enable = true;
+    capSysNice = true;
+  };
 
   # LACT
   services.lact.enable = true;
